requirements.txt
# Minimal dependencies (can be adjusted during implementation)
python-dotenv>=1.0.0
aiosqlite>=0.20.0
requests>=2.31.0
websockets>=12.0
telethon>=1.34.0
pydantic>=2.0
pyyaml>=6.0
watchdog>=4.0
pytest>=8.0
pytest-asyncio>=0.23
